add useExport
