+++ src/pages/hooks/home.hooks.ts
@@ -1,9 +1,9 @@
 import { useState, useCallback } from "react";
-import { writeTextFile, BaseDirectory } from "@tauri-apps/api/fs";
 import { PageMode, Account, AccountMap } from "../../types/";
 import { HomeState } from "./home.types";
 import { useImport } from "./useImport";
 import { useFilter } from "./useFilter";
+import { useExport } from "./useExport";
 
 export const useHome = (): HomeState => {
   const [mode, setMode] = useState<PageMode>("import");
@@ -30,50 +30,7 @@ export const useHome = (): HomeState => {
     accounts,
     updateAccountMap
   });
-
-  const convertAccountMapToCsv = useCallback(
-    (accountMap: AccountMap | null): string => {
-      // header: "url,username,source"
-      const header = "url,username,source";
-
-      // if accountMap is null, return header only
-      if (accountMap === null) {
-        return header;
-      }
-
-      // accountMap has the following structure:
-      // {
-      //   "icloud.com": [
-      //     {
-      //       username: "test",
-      //       source: "icloud"
-      //     }
-      //   ],...
-      // }
-      // convert to csv format like:
-      // "https://www.icloud.com,test,icloud"
-      const accountMapEntries = Object.entries(accountMap);
-      const accountMapCsv = accountMapEntries.reduce((acc, [key, value]) => {
-        const accountCsv = value.reduce((acc, account) => {
-          const { username, source } = account;
-          return `${acc}\n${key},${username},${source}`;
-        }, "");
-        return `${acc}${accountCsv}`;
-      }, header);
-      console.log(accountMapCsv.length);
-      return accountMapCsv;
-    },
-    []
-  );
-
-  const exportAccounts = useCallback(() => {
-    const fileContent = convertAccountMapToCsv(accountMap);
-    writeTextFile("export.csv", fileContent, {
-      dir: BaseDirectory.Download
-    })
-      .then(() => console.log("export ok"))
-      .catch((err) => console.log(err));
-  }, []);
+  const { exportAccounts } = useExport({ accountMap });
 
   return {
     mode,
