+++ src/pages/hooks/useExport/use-export.types.ts
@@ -1,0 +1,9 @@
+import { AccountMap } from "../../../types";
+
+export interface UseExportProps {
+  accountMap: AccountMap | null;
+}
+
+export interface ExportState {
+  exportAccounts: () => void;
+}
